requirements.txt
slack-sdk==3.29.0
slack-bolt==1.18.1
streamlit==1.31.0
openai==1.12.0
anthropic==0.18.1
ollama==0.1.7
python-dotenv==1.0.1
requests==2.31.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic

# Database dependencies
pg8000==1.31.2
pgvector==0.2.4

# Graph database
neo4j==5.16.0

# Embeddings and text processing
sentence-transformers==2.7.0
numpy==1.26.3
scikit-learn==1.4.0

# Text chunking and processing
beautifulsoup4==4.12.3